Major overhaul to DBTestCase design. Assumes that the test database does not already exist with populated schema, so it is up to individual test cases to specify and create a test database schema on demand. Significant additional testing needs to be done, as some of these changes are likely left incomplete, but this should illustrate the major elements.

+++ stride/clinical_item/psql/schemata/clinical_item.schema.sql
@@ -11,9 +11,9 @@ CREATE TABLE IF NOT EXISTS clinical_item
     name TEXT NOT NULL, -- e.g. MED540151
     description TEXT,   -- e.g. Warfarin
     default_recommend INTEGER DEFAULT 1, -- whether to recommend in CDSS
-    item_count INTEGER, -- summary statistic of # occurrences ()
-    patient_count INTEGER, -- summary statistic of # of unique patients
-    encounter_count INTEGER, -- summary statistic of # of unique encounters
+    item_count DOUBLE PRECISION, -- summary statistic of # occurrences ()
+    patient_count DOUBLE PRECISION, -- summary statistic of # of unique patients
+    encounter_count DOUBLE PRECISION, -- summary statistic of # of unique encounters
     analysis_status INTEGER DEFAULT 1, -- should be in assocation analysis?
     outcome_interest INTEGER DEFAULT 0, -- could be outcome measure?
     CONSTRAINT clinical_item_pkey PRIMARY KEY (clinical_item_id),
